Upgrade reactor to version 3 and fix robustness and stability issues. (#10)
* Remove Reduced operator

* Fix bug on isCoveredBy of RequestedDataProperties

* Modify CardinalityEstimator to use BigInteger

* Enhance ASK selector to incorporate graph

* Add cardinality estimation of Distinct and Reduced

* Fix NPE on ContextVar

* Upgrade to reactor 3.0.3

* Lift cost to use BigDecimal

* Use FilterOptimizer to push filters

* First impl of CrossProduct

* Fix cost estimation overflow

* Define CrossProduct cost approximation

* Fix cost estimation overflow

* Prune completePlans after finalization

* Calibrate cost estimation of SourceQuery

* Fix CrossProduct evaluation invokation

+++ core-api/src/main/java/org/semagrow/estimator/CardinalityEstimator.java
@@ -2,6 +2,8 @@ package org.semagrow.estimator;
 
 import org.eclipse.rdf4j.query.algebra.TupleExpr;
 
+import java.math.BigInteger;
+
 /**
  * Cardinality Estimator
  *
@@ -22,6 +24,6 @@ public interface CardinalityEstimator {
      * @param expr a logical or physical tree of expressions
      * @return an estimated number of results
      */
-    long getCardinality(TupleExpr expr);
+    BigInteger getCardinality(TupleExpr expr);
 
 }
